Check if listener is a callable before attaching
This change helps ensure that only a callable is attached to the
eventmanager, helping to eliminate objects that could never and should
never be attached. In a future commit I'm going to refactor the check to
be more meaningful, inclusive, and intuitive. But, for the time being,
this is what I'm going with.